ability.rb: allow user to :manage character; vs explicitly allowing actions die.rb: extracted Die from Roll; made it a non-db-backed model roll.rb: removed Die class; functionality remains (see die.rb) grids#show.html: warning about using browser zoom (and how it affects the drag/drop functionality) config/database.yml: added line for test db (sqlite3) grid.js: append tooltip to div instead of table cell; now we don't have to wait for the next full-refresh just to see the tooltip again, it stays with the token ability_spec: rspec testcase for permissions; WIP. roll_spec: rspec testcase for Roll; mostly complete; suggestions welcome.
NOTES:
You will need to add RSpec to the Gemfile or have the latest rspec/rspec-rails gem installed.

I plan on refactoring Ability.rb, so my aim is to make the spec as thorough as possible. This'll make the refactor [relatively] painless.

There is now a test database (RSpec won't run without it) configured.
I don't use fixtures currently; I may reconsider this; because I have some logic based around associations, which are a pain to mock/stub.

Also, grid.js previously said it was not licensed for redistribution; it was. Any version of grid.js stored on this github respoitory can be used under the MIT license. Not that you'd wanna use the old version anyways ;).